Implements control structure for motion detection and face detection and recognition
